+++ onnxruntime/core/providers/metawarenn/metawarenn_lib/op/constant.cc
@@ -4,7 +4,7 @@ namespace metawarenn {
 
 namespace op {
 
-Constant::Constant(std::string n_name, std::vector<int64_t> n_shape, std::vector<float> n_data, ElementType::element_type n_data_type)
+Constant::Constant(std::string n_name, std::vector<int> n_shape, std::vector<float> n_data, ElementType::element_type n_data_type)
          : Node(n_name, "Constant") {
   shape = n_shape;
   data = n_data;
