+++ onnxruntime/core/providers/metawarenn/metawarenn_lib/op/constant.h
@@ -9,8 +9,8 @@ namespace op {
 
 class Constant : public Node {
   public:
-    Constant(std::string n_name, std::vector<int64_t> n_shape, std::vector<float> n_data, ElementType::element_type n_data_type);
-    std::vector<int64_t> shape;
+    Constant(std::string n_name, std::vector<int> n_shape, std::vector<float> n_data, ElementType::element_type n_data_type);
+    std::vector<int> shape;
     std::vector<float> data;
     ElementType::element_type data_type;
 };
