+++ onnxruntime/core/providers/metawarenn/metawarenn_lib/op/input_data.cc
@@ -4,7 +4,7 @@ namespace metawarenn {
 
 namespace op {
 
-InputData::InputData(std::string n_name, std::vector<int64_t> n_shape, ElementType::element_type n_data_type)
+InputData::InputData(std::string n_name, std::vector<int> n_shape, ElementType::element_type n_data_type)
          : Node(n_name, "InputData") {
   shape = n_shape;
   data_type = n_data_type;
