+++ onnxruntime/core/providers/metawarenn/metawarenn_lib/op/input_data.h
@@ -9,8 +9,8 @@ namespace op {
 
 class InputData : public Node {
   public:
-    InputData(std::string n_name, std::vector<int64_t> n_shape, ElementType::element_type n_data_type);
-    std::vector<int64_t> shape;
+    InputData(std::string n_name, std::vector<int> n_shape, ElementType::element_type n_data_type);
+    std::vector<int> shape;
     ElementType::element_type data_type;
 };
